User controller finished, user service updated

+++ server/src/main/java/com/nocountry/server/service/impl/UserService.java
@@ -1,24 +1,29 @@
 package com.nocountry.server.service.impl;
 
+import com.nocountry.server.exception.UserNotFoundException;
 import com.nocountry.server.model.dto.UserDto;
 import com.nocountry.server.model.entity.User;
 import com.nocountry.server.repository.UserRepository;
 import com.nocountry.server.service.IUserService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class UserService implements IUserService {
 
-    @Autowired
-    private UserRepository repository;
-
+    private final UserRepository repository;
 
     @Override
     public User findById(Long id) {
-        return repository.findById(id).get();
+        Optional<User> user = repository.findById(id);
+        //if the user doesn't exist then generate an exception
+        return user.orElseThrow(()-> new UserNotFoundException("The user with that id doesn't exists") );
+
     }
 
     @Override
@@ -27,38 +32,53 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public void createUser(UserDto userDto) {
-        User user = new User();
-        user.setAccountLocked(false);
-        user.setEnabled(true);
-        user.setEmail(userDto.getEmail());
-        user.setLocation(userDto.getLocation());
-        user.setPassword(userDto.getPassword());
-        user.setFirstName(userDto.getFirstName());
-        user.setLastName(userDto.getLastName());
-        user.setUrlImage(userDto.getUrlImage());
+    public boolean createUser(UserDto userDto) {
+        if(!repository.existsByEmail(userDto.getEmail())){
+            User user = new User();
+            user.setAccountLocked(false);
+            user.setEnabled(true);
+            user.setEmail(userDto.getEmail());
+            user.setLocation(userDto.getLocation());
+            user.setPassword(userDto.getPassword());
+            user.setFirstName(userDto.getFirstName());
+            user.setLastName(userDto.getLastName());
+            user.setUrlImage(userDto.getUrlImage());
 
-        repository.save(user);
+            repository.save(user);
+            return true;
+        }
+            return false;
+
     }
 
     @Override
-    public void updateUser(UserDto userDto) {
+    public User updateUser(UserDto userDto, Long id) {
+        try{
+            User user = findById(id);
+            user.setAccountLocked(false);
+            user.setEnabled(true);
+            user.setEmail(userDto.getEmail());
+            user.setLocation(userDto.getLocation());
+            user.setPassword(userDto.getPassword());
+            user.setFirstName(userDto.getFirstName());
+            user.setLastName(userDto.getLastName());
+            user.setUrlImage(userDto.getUrlImage());
+            //updating the user
+            repository.save(user);
 
-        User user = new User();
-        user.setAccountLocked(false);
-        user.setEnabled(true);
-        user.setEmail(userDto.getEmail());
-        user.setLocation(userDto.getLocation());
-        user.setPassword(userDto.getPassword());
-        user.setFirstName(userDto.getFirstName());
-        user.setLastName(userDto.getLastName());
-        user.setUrlImage(userDto.getUrlImage());
-
-        repository.save(user);
+            return user;
+        }catch(UserNotFoundException ex){
+            //user doesn't exist, send null instead
+            return null;
+        }
     }
 
     @Override
-    public void deleteUser(Long id) {
-        repository.deleteById(id);
+    public boolean deleteUser(Long id) {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
